Tensor deallocation through deleter functions.
Now, tensors can register their deleter functions with
ExecutionContext. ExecutionContext iterates through deleters
and calls them one by one. The deleters are implemented by
tensor implementations, and can be  accessed by
ExecutionContext which is a fried of TensorBase now.

The current design is not thread safe.  Each deleter stores a shared
pointer to the underlying resource, and it only frees the resource if it
is the last owner.


Former-commit-id: 987564b5319200c6a47866789e044b2754d3d3d4
Former-commit-id: aa05fea0b1a131f1cfa55fc796efb655be80d8a3
Former-commit-id: 51564aaa7b4d6b31b50de8a5d8246378d936b298
Former-commit-id: bde1de89bb7a3d8d955f1c80212c1d3e9b6cdfd9 [formerly 2a89f2acbc1731f1350c2bf01d5c50b99a8583ee]
Former-commit-id: 0c46410baec1a94ee7120415a9912eb246fc3cb0