Tensor deallocation through deleter functions.
Now, tensors can register their deleter functions with
ExecutionContext. ExecutionContext iterates through deleters
and calls them one by one. The deleters are implemented by
tensor implementations, and can be  accessed by
ExecutionContext which is a fried of TensorBase now.

The current design is not thread safe.  Each deleter stores a shared
pointer to the underlying resource, and it only frees the resource if it
is the last owner.