Fixed double-delete for the default runtime engine.
The execution context was copying the pointer for the default
runtime engine. When multiple copies were made, that pointer would
get deleted twice. We provided an explicit copy constructor and
assignment to make sure we make a new object that can be safely
deleted in a copy of the execution context.

For the future: should we be copying execution context and runtime
engine?


Former-commit-id: 45f1c3519501716bece300f975debdbd0fa9638b