Fixed double-delete for the default runtime engine.
The execution context was copying the pointer for the default
runtime engine. When multiple copies were made, that pointer would
get deleted twice. We provided an explicit copy constructor and
assignment to make sure we make a new object that can be safely
deleted in a copy of the execution context.

For the future: should we be copying execution context and runtime
engine?


Former-commit-id: 9c6d7988736cbab00e9ef399b5dba2b42d75e232
Former-commit-id: 0bb48b482ec1419bbe8f1f84656ba1897242385f
Former-commit-id: 003274333d861e10fd1ba567462a447ab56ada4e
Former-commit-id: cb1e40c80b5914efa5db5aefe627d2d10faf8cc4 [formerly bc6f89711cc4ecc5a2b9a09ae3accda093814ff0]
Former-commit-id: b51a0b80853931c966581df41b80c69b32e64ee7